First pass at util functions for loading data

+++ src/predecl/GeometricDistrDecl.cpp
@@ -27,7 +27,7 @@ std::shared_ptr<ir::Expr> GeometricDistrDecl::getNew(
 
   auto ret = std::make_shared<ir::Distribution>(this->getName(), this);
   ret->setArgs(args);
-  ret->setTyp(fact->getTy(ir::IRConstString::INT));
+  ret->setTyp(fact->getTy(ir::IRConstString::DOUBLE));
   ret->processArgRandomness();
   ret->setRandom(true);
   return ret;
